// ts/enums.d.ts
export const enum CardType {
	MINION = 4,
	INVALID = 0,
	HERO = 3,
	TOKEN = 9,
	SPELL = 5,
	ENCHANTMENT = 6,
	HERO_POWER = 10,
	ABILITY = 5,
	PLAYER = 2,
	ITEM = 8,
	GAME = 1,
	WEAPON = 7
}

export const enum PlayState {
	INVALID = 0,
	CONCEDED = 8,
	DISCONNECTED = 7,
	LOSING = 3,
	WINNING = 2,
	TIED = 6,
	LOST = 5,
	WON = 4,
	QUIT = 8,
	PLAYING = 1
}

export const enum ChoiceType {
	INVALID = 0,
	MULLIGAN = 1,
	GENERAL = 2
}
// ...
export const enum Mulligan {
	INPUT = 1,
	INVALID = 0,
	DEALING = 2,
	WAITING = 3,
	DONE = 4
}
// ...
export const enum BlockType {
	DEATHS = 6,
	POWER = 3,
	ACTION = 99,
	SCRIPT = 4,
	ATTACK = 1,
	JOUST = 2,
	TRIGGER = 5,
	CONTINUOUS = 2,
	PLAY = 7,
	FATIGUE = 8,
	RITUAL = 9
}

export const enum FormatType {
	FT_UNKNOWN = 0,
	FT_WILD = 1,
	FT_STANDARD = 2
}

export const enum Rarity {
	LEGENDARY = 5,
	EPIC = 4,
	INVALID = 0,
	UNKNOWN_6 = 6,
	RARE = 3,
	COMMON = 1,
	FREE = 2
}

export const enum Zone {
	DISCARD = -2,
	INVALID = 0,
	SETASIDE = 6,
	HAND = 3,
	DECK = 2,
	REMOVEDFROMGAME = 5,
	PLAY = 1,
	GRAVEYARD = 4,
	SECRET = 7
}